refactor: DAO 추상화
