+++ src/main/java/subway/dao/LineDao.java
@@ -1,62 +1,18 @@
 package subway.dao;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import javax.sql.DataSource;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 import subway.entity.LineEntity;
 
-@Repository
-public class LineDao {
+import java.util.List;
 
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert insertAction;
+public interface LineDao {
 
-    private final RowMapper<LineEntity> rowMapper = (rs, rowNum) ->
-            new LineEntity(
-                    rs.getLong("id"),
-                    rs.getString("name"),
-                    rs.getString("color")
-            );
+    LineEntity insert(LineEntity lineEntity);
 
-    public LineDao(JdbcTemplate jdbcTemplate, DataSource dataSource) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.insertAction = new SimpleJdbcInsert(dataSource)
-                .withTableName("line")
-                .usingGeneratedKeyColumns("id");
-    }
+    List<LineEntity> findAll();
 
-    public LineEntity insert(LineEntity lineEntity) {
-        Map<String, Object> params = new HashMap<>();
-        params.put("id", lineEntity.getId());
-        params.put("name", lineEntity.getName());
-        params.put("color", lineEntity.getColor());
+    LineEntity findById(Long id);
 
-        Long lineId = insertAction.executeAndReturnKey(params).longValue();
-        return new LineEntity(lineId, lineEntity.getName(), lineEntity.getColor());
-    }
+    void update(LineEntity newLineEntity);
 
-    public List<LineEntity> findAll() {
-        String sql = "SELECT id, name, color FROM line";
-        return jdbcTemplate.query(sql, rowMapper);
-    }
-
-    public LineEntity findById(Long id) {
-        String sql = "SELECT id, name, color FROM line WHERE id = ?";
-        return jdbcTemplate.queryForObject(sql, rowMapper, id);
-    }
-
-    public void update(LineEntity newLineEntity) {
-        String sql = "UPDATE line SET name = ?, color = ? WHERE id = ?";
-        jdbcTemplate.update(sql,
-                new Object[]{newLineEntity.getName(), newLineEntity.getColor(), newLineEntity.getId()});
-    }
-
-    public void deleteById(Long id) {
-        jdbcTemplate.update("DELETE FROM line WHERE id = ?", id);
-    }
+    void deleteById(Long id);
 }
