+++ src/main/java/subway/dao/StationDao.java
@@ -1,58 +1,18 @@
 package subway.dao;
 
-import java.util.List;
-import javax.sql.DataSource;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 import subway.entity.StationEntity;
 
-@Repository
-public class StationDao {
+import java.util.List;
 
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert insertAction;
+public interface StationDao {
 
-    private final RowMapper<StationEntity> rowMapper = (rs, rowNum) ->
-            new StationEntity(
-                    rs.getLong("id"),
-                    rs.getString("name")
-            );
+    StationEntity insert(StationEntity stationEntity);
 
+    List<StationEntity> findAll();
 
-    public StationDao(JdbcTemplate jdbcTemplate, DataSource dataSource) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.insertAction = new SimpleJdbcInsert(dataSource)
-                .withTableName("station")
-                .usingGeneratedKeyColumns("id");
-    }
+    StationEntity findById(Long id);
 
-    public StationEntity insert(StationEntity stationEntity) {
-        SqlParameterSource params = new BeanPropertySqlParameterSource(stationEntity);
-        Long id = insertAction.executeAndReturnKey(params).longValue();
-        return new StationEntity(id, stationEntity.getName());
-    }
+    void update(StationEntity newStationEntity);
 
-    public List<StationEntity> findAll() {
-        String sql = "SELECT * FROM station";
-        return jdbcTemplate.query(sql, rowMapper);
-    }
-
-    public StationEntity findById(Long id) {
-        String sql = "SELECT * FROM station WHERE id = ?";
-        return jdbcTemplate.queryForObject(sql, rowMapper, id);
-    }
-
-    public void update(StationEntity newStationEntity) {
-        String sql = "UPDATE station SET name = ? WHERE id = ?";
-        jdbcTemplate.update(sql, new Object[]{newStationEntity.getName(), newStationEntity.getId()});
-    }
-
-    public void deleteById(Long id) {
-        String sql = "DELETE FROM STATION WHERE id = ?";
-        jdbcTemplate.update(sql, id);
-    }
+    void deleteById(Long id);
 }
