+++ src/main/java/subway/service/LineService.java
@@ -1,7 +1,5 @@
 package subway.service;
 
-import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import subway.dao.LineDao;
 import subway.domain.Line;
@@ -10,6 +8,9 @@ import subway.dto.response.LineResponse;
 import subway.entity.LineEntity;
 import subway.mapper.LineMapper;
 import subway.repository.LineRepository;
+
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class LineService {
